\name{iris}
\alias{iris}
\docType{data}
\title{Iris Dataset}
\description{
  The iris dataset consists of 150 observations of five variables, describing the length and width of sepals and petals for three iris species.
}
\usage{data("iris")}
\source{\url{https://archive.ics.uci.edu}}
\references{
  Fisher, R. A. (1988). Iris. UCI Machine Learning Repository. \url{https://doi.org/10.24432/C56C76}.
}
\keyword{datasets}
