\name{iris_uci}
\alias{iris_uci}
\docType{data}
\title{iris dataset}
\description{
  The iris dataset consists of 150 observations of five variables, describing the length and width of sepals and petals for three iris species.
}
\usage{data("iris_uci")}
\source{\url{https://archive.ics.uci.edu}}
\references{
  Fisher, R. A. (1988). Iris. UCI Machine Learning Repository. \url{doi.org/10.24432/C56C76}.
}
\keyword{datasets}
