% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca}
\alias{pca}
\title{Plot the Principal Component Analysis}
\usage{
pca(data, groups = "none")
}
\arguments{
\item{data}{Dataset in \code{data.frame} format.}

\item{groups}{This parameter allows coloring of observations according
to their group and drawing ellipses around each group with a confidence
level of 0.98. The default is 'none'.}
}
\value{
A \code{ggplot}.
}
\description{
Apply PCA (Principal Component Analysis) to the data and plot a scatterplot of
the first two principal components.
}
\examples{

# Load the packages
library(hclusteasy)


# Read iris dataset from packege
data("iris_uci")

# Select column "Species" (groups) from the iris dataset
species <- iris_uci[, 5]

# Remove column "Species" from the iris dataset
iris <- iris_uci[, -5]


# Apply pca and ploting the two firsts components in pca, without groups
pca(iris)

# Apply pca and ploting the first two components in pca, with groups
pca(iris, groups = species)

}
