% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.data.R
\name{read.data}
\alias{read.data}
\title{Read Dataset in txt, xlsx, or xls Format}
\usage{
read.data(path, col.names = FALSE, col.types = NULL)
}
\arguments{
\item{path}{The path to the folder where the dataset is located.}

\item{col.names}{\code{Logical} value indicating whether the first row
of the dataset should be used as column names. Default is \code{FALSE}.}

\item{col.types}{Character or a character vector specifying
the data types for each column: "skip", "guess", "logical", "numeric",
"date", "text", or "list". By default, it is \code{NULL}, which means the
data types will be determined automatically. Note that \code{txt} files do not
support parameter \code{col.types}.}
}
\value{
A dataset in \code{data.frame} format.
}
\description{
Read a dataset from a specified file path in
\code{txt}(separated by space), \code{xlsx}, or \code{xls} format and return a \code{data.frame}.
}
\examples{
# Load the package
library(hclusteasy)

# Set the file path
file_path <- system.file("extdata",
                          "iris_uci.xlsx",
                           package = "hclusteasy")


# Read a XLSX dataset
iris <- read.data(file_path,col.names = TRUE)

}
