\name{wine_uci}
\alias{wine_uci}
\docType{data}
\title{wine dataset}
\description{
  These data are the results of a chemical analysis of wines grown in the same region in Italy but derived from three different cultivars. The analysis determined the quantities of 13 constituents found in each of the three types of wines.
}
\usage{data("wine_uci")}
\source{\url{https://archive.ics.uci.edu}}
\references{
  Aeberhard, Stefan and Forina, M. (1991). Wine. UCI Machine Learning Repository. \url{https://doi.org/10.24432/C5PC7J}.
}
\keyword{datasets}
