\name{iris_uci}
\alias{iris_uci}
\docType{data}
\title{Iris Dataset}
\description{
  This dataset contains 150 flower samples distributed among 3 iris species classes: Setosa, Versicolor, and Virginica. It consists of 5 columns, including 4 attributes measured in centimeters: sepal length and width, and petal length and width, along with a column indicating the iris species. This dataset was introduced by Ronald A. Fisher in 1936 in his classic paper on linear discriminant analysis.
}
\usage{data("iris_uci")}
\source{\url{https://archive.ics.uci.edu}}
\references{
  Fisher, R. A. (1988). Iris. UCI Machine Learning Repository. \doi{10.24432/C56C76}.
}
\keyword{datasets}
