\name{wine_uci}
\alias{wine_uci}
\docType{data}
\title{Wine Dataset}
\description{
  It consists of a dataset containing 178 wine samples distributed into 3 distinct classes. It has 14 columns, comprising 13 chemical attributes such as alcohol content, malic acid amount, ash, alkalinity of ash, magnesium, phenols, flavonoids, proanthocyanins, color intensity, hue, OD280/OD315 ratio, and proline, along with one column indicating the wine class. This dataset was introduced by Forina et al. in 1991 in a study on the chemical analysis of wines grown in the regions of Italy.
}
\usage{data("wine_uci")}
\source{\url{https://archive.ics.uci.edu}}
\references{
  Aeberhard, Stefan and Forina, M. (1991). Wine. UCI Machine Learning Repository. \doi{10.24432/C5PC7J}.
}
\keyword{datasets}
